requirement.txt
streamlit
numpy
pillow
ultralytics
